feat(BetterScreenshare): reduce maximum video bitrate
I made a sad discovery that Discord has some kind of mechanism that limits the video bitrate to 10000 Kb/s. (It still looks a lot better). If you exceed the limit, everyone will experience packet loss for some reason regardless of upload speed. Maybe there will be a workaround in the future.