Fix(Uwuify): Fixed the corruption of URLS
did essentially a whole rewrite of the Uwuifier to operate word-wise
instead of upon the entire message using regexes. Also retains
capitalization now, which may or may not actually be wanted.

As a result of word-wise operations, it can now avoid uwuifying certain
patterns, currently only does so for URLS

resolves #1186