feat(MessageLogger): add option to ignore channels
Add an option to outright ignore channels from creating logs with the
MessageLogger plugin.

In some high-traffic channels that receive a lot of edits and/or
deletions (by virtue of so-called "sticky messages"), the MessageLogger
plugin tends to get in the way of reading the chat history.

This patch adds an option to provide a comma-separated list of channel
IDs to ignore from logging.

Local tests confirm it is working as intended. To the best of my
knowledge, I followed the style guide of this project.
However, as this is my first time working with TypeScript and the
Vencord codebase, please feel free to give feedback and pointers on how
to proceed, so we can get this merged.

Unfortunately, I was not able to find any relevant discussions or issues
with regards to such a feature request. If there are any, please do link
them.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>