Fix command API
Fix the command API, broken by discord removing the /giphy command, which was used to locate the injection point for the CommandAPI _init() function. This fix is only a sketchy temporary fix and i am uncertain whether it should be used over a better, more stable solution.. but it works for now for me ig? I am mostly worried that i have seen some users who still have /giphy.. so this would break  the CommandAPI for them..