Delete FixImagesQuality
Discord fixed their issue with images getting converted to webp losing quality. In the future this plugin can be re-purposed to keep images format as the original, but that requires more than just removing part of the Discord code, and a bit of study to make sure it does not cause issues.