MessageLinkEmbeds: Add option to only load visible links
If a channel has many message links, MessageLinkEmbeds may load many
link embeds, even for messages that are not in the user's viewport. As
of writing, the plugin loads embeds from the oldest message to the
newest message, so the embeds for the most recent messages do not appear
until all the old embeds are loaded.

Add an option to only load link embeds in the user's viewport to work
around this issue. This will result in less API requests and faster
loading for channels with many message links, but may cause embeds to
suddenly appear when scrolling through a channel, so this option is off by
default.