Make context menu patch not mutate props
This makes the menu hooks pure, which allows React hooks to be used.
Eventually the stuff about returning callbacks may be phased out in favor of useLayout.

This approach is a bit heavy-handed however, since it deeply clones the entire props.
Gonna try to make a more fine-grained cloner later.