Fix: PetPet Plugin's Slash Command Image Option
The image option for the PetPet command has not been working for about 5 months now, as noted in the closed #2070 pull request. This finally fixes it so it actually works again. This also includes clearing the slash command draft after you run it, so if you run the slash command again, the previously uploaded file from the previous slash command run will not be there in the upload area. Also made DraftType use webpack common and implemented it into the _fakeNitro_ and _crashHandler_ plugins. Along with DraftType, I also added UploadModule (which is the thing that allows draft clear to happen) to webpack common to make it easier for future use.