VcNarrator: Make` {{NICKNAME}}` placeholder fall back to display name before username
This commit is a proposed amendment of #1792.

Currently, the `{{NICKNAME}}` placeholder allows reading a user's nickname in voice events. If the user has no nickname, it uses the username instead.

I would like to propose the user's global display name be used, if available, before falling back to the username, because the username is always all lowercase, often difficult for the text-to-speech system to pronounce, and may not reflect how the user prefers to be known.

This commit makes that change.