Reactiveness of Components & float support for NumberSetting
Current Plugin Settings Components are not reactive whenever I attempt to change the state of the components programatically (e.g. when I need to reset Plugin's settings from itself). This commit aims to address this issue, helping to refresh the UI of individual components in PluginModal without need of closing the entire Modal and reopening it.

In this commit, NumberSetting component is also getting more robust handling of `Number` datatype. Original file was assuming an Integer to be strictly on input (otherwise crashed), however Number type can be also a decimal (floating point) number.