feat(MessageLoggerEnhanced): Add optional time-based cleanup (#603)
* feat(MessageLoggerEnhanced): Add optional time-based cleanup
- temporal cleanup preserves DMs and messages in the current channel by default
- proper dm detection via ChannelStore (lmk if this is the wrong way to check)
- add time limit and server message preservation to mle settings

* feat(EquicordHelper): add feat of removed vencord plugin NoDefaultHangStatus
it anchors to the default hang status and force it
to set enableHangStatus to false. im not sure if it
really belongs in EquicordHelper but it feels too
small to be its own plugin, and it doesnt fit anywhere
else that I can see. Let me know what you think.